+++ portals/developer-portal/src/controllers/apiContentController.js
@@ -25,8 +25,11 @@ const path = require('path');
 const exphbs = require('express-handlebars');
 const util = require('../utils/util');
 const constants = require('../utils/constants');
-const adminDao = require('../dao/adminDao');
-const apiDao = require('../dao/apiMetadataDao');
+const orgDao = require('../dao/organizationDao');
+const appDao = require('../dao/applicationDao');
+const apiDao = require('../dao/apiDao');
+const apiFileDao = require('../dao/apiFileDao');
+const viewDao = require('../dao/viewDao');
 const subDao = require('../dao/subscriptionDao');
 const apiMetadataService = require('../services/apiMetadataService');
 const { apiUsesApiKeySecurity, findSubscriptionTokenHeader } = require('../utils/apiDefinitionUtil');
@@ -59,7 +62,7 @@ const loadAPIs = async (req, res) => {
         const listingPage = isMcpListing ? 'pages/mcp' : 'pages/apis';
         html = renderTemplate(layoutPath + listingPage + '/page.hbs', layoutPath + 'layout/main.hbs', templateContent, false);
     } else {
-        const orgDetails = await adminDao.getOrganization(orgName);
+        const orgDetails = await orgDao.get(orgName);
         const devportalMode = orgDetails.ORG_CONFIG?.devportalMode || constants.DEVPORTAL_MODE.DEFAULT;
         try {
             const orgID = orgDetails.ORG_ID;
@@ -81,7 +84,7 @@ const loadAPIs = async (req, res) => {
 
             // Fetch all applications for the user/org once
             if (req.user) {
-                allApplications = await adminDao.getApplications(orgID, req.user.sub);
+                allApplications = await appDao.list(orgID, req.user.sub);
             }
             for (const metaData of metaDataList) {
                 metaData.subscriptionPolicyDetails = await util.appendSubscriptionPlanDetails(orgID, metaData.subscriptionPolicies);
@@ -89,7 +92,7 @@ const loadAPIs = async (req, res) => {
                 if (allApplications.length > 0) {
                     perApiAppList = await Promise.all(
                         allApplications.map(async (app) => {
-                            const subscription = await adminDao.getAppApiSubscription(orgID, app.APP_ID, metaData.apiID);
+                            const subscription = await subDao.getByAppAndApi(orgID, app.APP_ID, metaData.apiID);
                             const subscriptionData = subscription.length > 0 ? {
                                 policyId: subscription[0].POLICY_ID,
                                 policyName: metaData.subscriptionPolicies.find(p => p.policyID === subscription[0].POLICY_ID)?.policyName || 'Unknown'
@@ -108,7 +111,7 @@ const loadAPIs = async (req, res) => {
             // Load subscriptions for APIs with subscription plans (single call for all)
             if (req.user) {
                 try {
-                    const localSubs = await subDao.listSubscriptions(orgID);
+                    const localSubs = await subDao.list(orgID);
                     const subscribedApiIds = new Set(localSubs.map(sub => sub.API_ID));
                     for (const metaData of metaDataList) {
                         const hasPlans = (metaData.subscriptionPolicies || []).length > 0;
@@ -238,12 +241,12 @@ const loadAPIContent = async (req, res) => {
         html = renderTemplate(layoutPath + landingPage + '/page.hbs', layoutPath + 'layout/main.hbs', templateContent, false);
         res.send(html);
     } else {
-        const orgDetails = await adminDao.getOrganization(orgName);
+        const orgDetails = await orgDao.get(orgName);
         const devportalMode = orgDetails.ORG_CONFIG?.devportalMode || constants.DEVPORTAL_MODE.DEFAULT;
         try {
-            const orgDetails = await adminDao.getOrganization(orgName);
+            const orgDetails = await orgDao.get(orgName);
             const orgID = orgDetails.ORG_ID;
-            const apiID = await apiDao.getAPIId(orgID, apiHandle);
+            const apiID = await apiDao.getId(orgID, apiHandle);
             const metaData = await loadAPIMetaData(req, orgID, apiID);
             
             // Log API access for audit trail
@@ -269,9 +272,9 @@ const loadAPIContent = async (req, res) => {
             let apiDetails = "";
             let schemaDefinition = "";
             let apiDefinition = {};
-            const markdownResponse = await apiDao.getAPIFile(constants.FILE_NAME.API_MD_CONTENT_FILE_NAME, constants.DOC_TYPES.API_LANDING, orgID, apiID);
+            const markdownResponse = await apiFileDao.get(constants.FILE_NAME.API_MD_CONTENT_FILE_NAME, constants.DOC_TYPES.API_LANDING, orgID, apiID);
             if (!markdownResponse) {
-                let additionalAPIContentResponse = await apiDao.getAPIFile(constants.FILE_NAME.API_HBS_CONTENT_FILE_NAME, constants.DOC_TYPES.API_LANDING, orgID, apiID);
+                let additionalAPIContentResponse = await apiFileDao.get(constants.FILE_NAME.API_HBS_CONTENT_FILE_NAME, constants.DOC_TYPES.API_LANDING, orgID, apiID);
                 if (!additionalAPIContentResponse) {
                     loadDefault = true;
                     if (
@@ -306,7 +309,7 @@ const loadAPIContent = async (req, res) => {
                     }
                     if (metaData.apiInfo.apiType === constants.API_TYPE.GRAPHQL) {
                         apiDefinition = "";
-                        apiDefinition = await apiDao.getAPIFile(constants.FILE_NAME.API_DEFINITION_GRAPHQL, constants.DOC_TYPES.API_DEFINITION, orgID, apiID);
+                        apiDefinition = await apiFileDao.get(constants.FILE_NAME.API_DEFINITION_GRAPHQL, constants.DOC_TYPES.API_DEFINITION, orgID, apiID);
                         apiDefinition = apiDefinition.API_FILE.toString(constants.CHARSET_UTF8);
                         apiDetails = {
                             title: metaData.apiInfo.apiName || "No title",
@@ -325,7 +328,7 @@ const loadAPIContent = async (req, res) => {
                         apiDetails = {};
                         apiDetails['serverDetails'] = mcpProductionURL ? { productionURL: mcpProductionURL, sandboxURL: '' } : '';
                         try {
-                            let rawSchema = await apiDao.getAPIFileByType(
+                            let rawSchema = await apiFileDao.getByType(
                                 constants.DOC_TYPES.SCHEMA_DEFINITION,
                                 orgID,
                                 apiID
@@ -362,11 +365,11 @@ const loadAPIContent = async (req, res) => {
             }
             let appList = [];
             if (req.user) {
-                let applications = await adminDao.getApplications(orgID, req.user.sub);
+                let applications = await appDao.list(orgID, req.user.sub);
                 if (applications.length > 0) {
                     appList = await Promise.all(
                         applications.map(async (app) => {
-                            const subscription = await adminDao.getAppApiSubscription(orgID, app.APP_ID, metaData.apiID);
+                            const subscription = await subDao.getByAppAndApi(orgID, app.APP_ID, metaData.apiID);
                             const subscriptionData = subscription.length > 0 ? {
                                 policyId: subscription[0].POLICY_ID,
                                 policyName: metaData.subscriptionPolicies.find(p => p.policyID === subscription[0].POLICY_ID)?.policyName || 'Unknown'
@@ -386,7 +389,7 @@ const loadAPIContent = async (req, res) => {
             const hasPlans = (subscriptionPlans || []).length > 0;
             if (req.user && hasPlans) {
                 try {
-                    const localSubs = await subDao.listSubscriptions(orgID, { apiId: apiID });
+                    const localSubs = await subDao.list(orgID, { apiId: apiID });
                     subscriptions = (localSubs || []).map(sub => ({
                         subscriptionId: sub.SUB_ID,
                         subscriptionPlanName: sub.DP_SUBSCRIPTION_POLICY?.DISPLAY_NAME || sub.DP_SUBSCRIPTION_POLICY?.POLICY_NAME || '',
@@ -496,8 +499,8 @@ const getAPIDefinition = async (orgName, viewName, apiHandle) => {
             }
         }
     } else {
-        const orgID = await adminDao.getOrgId(orgName);
-        const apiID = await apiDao.getAPIId(orgID, apiHandle);
+        const orgID = await orgDao.getId(orgName);
+        const apiID = await apiDao.getId(orgID, apiHandle);
         metaData = await apiMetadataService.getMetadataFromDB(orgID, apiID, viewName);
         const data = metaData ? JSON.stringify(metaData) : {};
         metaData = JSON.parse(data);
@@ -509,7 +512,7 @@ const getAPIDefinition = async (orgName, viewName, apiHandle) => {
             templateContent.swagger = JSON.stringify({ servers: [{ url: productionURL }] });
             // Load MCP schema so loadAPIDefinitionRaw can serve it via SPEC_FORMAT_MAP field:'schema'
             try {
-                const rawSchema = await apiDao.getAPIFileByType(constants.DOC_TYPES.SCHEMA_DEFINITION, orgID, apiID);
+                const rawSchema = await apiFileDao.getByType(constants.DOC_TYPES.SCHEMA_DEFINITION, orgID, apiID);
                 if (rawSchema?.API_FILE) {
                     templateContent.schema = rawSchema.API_FILE.toString(constants.CHARSET_UTF8);
                 }
@@ -517,7 +520,7 @@ const getAPIDefinition = async (orgName, viewName, apiHandle) => {
                 logger.warn('Could not load MCP schema definition for raw spec', { orgID, apiID, error: schemaErr.message });
             }
         } else if (metaData.apiInfo.apiType === constants.API_TYPE.GRAPHQL) {
-            apiDefinition = await apiDao.getAPIFile(constants.FILE_NAME.API_DEFINITION_GRAPHQL, constants.DOC_TYPES.API_DEFINITION, orgID, apiID);
+            apiDefinition = await apiFileDao.get(constants.FILE_NAME.API_DEFINITION_GRAPHQL, constants.DOC_TYPES.API_DEFINITION, orgID, apiID);
             templateContent.graphql = apiDefinition.API_FILE.toString(constants.CHARSET_UTF8);
         } else {
             apiDefinition = await getApiDefinitionFileContent(orgID, apiID);
@@ -553,12 +556,12 @@ const loadDocsPage = async (req, res) => {
         }
         html = renderTemplate(layoutPath + 'pages/docs/page.hbs', layoutPath + 'layout/main.hbs', templateContent, false);
     } else {
-        const orgDetails = await adminDao.getOrganization(orgName);
+        const orgDetails = await orgDao.get(orgName);
         const devportalMode = orgDetails.ORG_CONFIG?.devportalMode || constants.DEVPORTAL_MODE.DEFAULT;
 
         try {
-            const orgID = await adminDao.getOrgId(orgName);
-            const apiID = await apiDao.getAPIId(orgID, apiHandle);
+            const orgID = await orgDao.getId(orgName);
+            const apiID = await apiDao.getId(orgID, apiHandle);
             const viewName = req.params.viewName;
             const docNames = await apiMetadataService.getAPIDocTypes(orgID, apiID);
 
@@ -573,7 +576,7 @@ const loadDocsPage = async (req, res) => {
                 }
             }
 
-            const apiMetadata = await apiDao.getAPIMetadata(orgID, apiID);
+            const apiMetadata = await apiDao.get(orgID, apiID);
             let apiType = apiMetadata[0].dataValues.API_TYPE;
             const gatewayType = apiMetadata[0].dataValues.GATEWAY_TYPE;
             const metaForNav = {
@@ -672,7 +675,7 @@ const loadDocument = async (req, res) => {
         return;
     }
 
-    const orgDetails = await adminDao.getOrganization(orgName);
+    const orgDetails = await orgDao.get(orgName);
     const devportalMode = orgDetails.ORG_CONFIG?.devportalMode || constants.DEVPORTAL_MODE.DEFAULT;
     let baseDocUrl = '/' + orgName + '/views/' + viewName + "/api/" + apiHandle
     if (req.originalUrl.includes('/mcp')) {
@@ -779,11 +782,11 @@ const loadDocument = async (req, res) => {
             templateContent.isAPIDefinition = true;
         }
         try {
-            const orgID = await adminDao.getOrgId(orgName);
-            const apiID = await apiDao.getAPIId(orgID, apiHandle);
+            const orgID = await orgDao.getId(orgName);
+            const apiID = await apiDao.getId(orgID, apiHandle);
             const viewName = req.params.viewName;
             let docNames = await apiMetadataService.getAPIDocTypes(orgID, apiID);
-            const apiMetadata = await apiDao.getAPIMetadata(orgID, apiID);
+            const apiMetadata = await apiDao.get(orgID, apiID);
             let apiType = apiMetadata[0].dataValues.API_TYPE;
             const referenceID = apiMetadata[0].dataValues.REFERENCE_ID;
             // All MCPs (registry and CP) need a Specification entry in the sidebar
@@ -918,7 +921,7 @@ function resolveSamplesPath(apiHandle) {
 }
 
 async function getApiDefinitionFileContent(orgID, apiID) {
-    const apiDefinition = await apiDao.getAPIDoc(constants.DOC_TYPES.API_DEFINITION, orgID, apiID);
+    const apiDefinition = await apiFileDao.getDoc(constants.DOC_TYPES.API_DEFINITION, orgID, apiID);
     if (apiDefinition?.API_FILE) {
         return apiDefinition.API_FILE.toString(constants.CHARSET_UTF8);
     }
@@ -1111,14 +1114,14 @@ const loadAPIContentMd = async (req, res) => {
     const { orgName, apiHandle, viewName } = req.params;
 
     try {
-        const orgDetails = await adminDao.getOrganization(orgName);
+        const orgDetails = await orgDao.get(orgName);
         const orgID = orgDetails.ORG_ID;
 
         if (await isAiDisabledForPortal(orgID, viewName)) {
             return res.status(404).send('# Not Found\n\nThis resource is not available for agents.');
         }
 
-        const apiID = await apiDao.getAPIId(orgID, apiHandle);
+        const apiID = await apiDao.getId(orgID, apiHandle);
         const metaData = await loadAPIMetaData(req, orgID, apiID);
 
         if (metaData.apiInfo?.agentVisibility === 'HIDDEN') {
@@ -1139,16 +1142,16 @@ const loadAPIContentMd = async (req, res) => {
         try {
             if (apiType === constants.API_TYPE.GRAPHQL) {
                 specHeading = 'GraphQL Schema';
-                const raw = await apiDao.getAPIFile(constants.FILE_NAME.API_DEFINITION_GRAPHQL, constants.DOC_TYPES.API_DEFINITION, orgID, apiID);
+                const raw = await apiFileDao.get(constants.FILE_NAME.API_DEFINITION_GRAPHQL, constants.DOC_TYPES.API_DEFINITION, orgID, apiID);
                 if (raw) apiDefinition = raw.API_FILE.toString(constants.CHARSET_UTF8);
             } else if (apiType === constants.API_TYPE.MCP) {
                 specHeading = 'Tool Schema';
-                const raw = await apiDao.getAPIDoc(constants.DOC_TYPES.SCHEMA_DEFINITION, orgID, apiID, null);
+                const raw = await apiFileDao.getDoc(constants.DOC_TYPES.SCHEMA_DEFINITION, orgID, apiID, null);
                 if (raw) apiDefinition = raw.API_FILE.toString(constants.CHARSET_UTF8);
             } else {
                 if (apiType === constants.API_TYPE.WS || apiType === constants.API_TYPE.WEBSUB) specHeading = 'AsyncAPI Specification';
                 else if (apiType === 'SOAP') specHeading = 'WSDL';
-                const raw = await apiDao.getAPIFile(constants.FILE_NAME.API_DEFINITION_FILE_NAME, constants.DOC_TYPES.API_DEFINITION, orgID, apiID);
+                const raw = await apiFileDao.get(constants.FILE_NAME.API_DEFINITION_FILE_NAME, constants.DOC_TYPES.API_DEFINITION, orgID, apiID);
                 if (raw) apiDefinition = raw.API_FILE.toString(constants.CHARSET_UTF8);
             }
         } catch (defErr) {
@@ -1246,8 +1249,7 @@ async function buildLlmsTxtTemplateContent(req, orgID, orgName, viewName, config
         if (byType[type]) byType[type].push(api);
     }
 
-    const apiMetadataDao = require('../dao/apiMetadataDao');
-    const viewId = await apiMetadataDao.getViewID(orgID, viewName);
+    const viewId = await viewDao.getId(orgID, viewName);
     const allApiFlows = await apiFlowService.getAllAPIFlowsFromDB(orgID, viewId);
     const allPublishedWorkflows = allApiFlows.filter(flow => flow.status === 'PUBLISHED');
     const publishedWorkflows = allPublishedWorkflows.filter(flow => flow.agentVisibility !== 'HIDDEN');
@@ -1275,10 +1277,10 @@ async function buildLlmsTxtTemplateContent(req, orgID, orgName, viewName, config
 const loadLlmsTxt = async (req, res) => {
     const { orgName, viewName } = req.params;
     try {
-        const orgDetails = await adminDao.getOrganization(orgName);
+        const orgDetails = await orgDao.get(orgName);
         const orgID = orgDetails.ORG_ID;
 
-        const configAsset = await adminDao.getOrgContent({
+        const configAsset = await orgDao.getContent({
             orgId: orgID, fileType: constants.FILE_TYPE.LLMS_CONFIG, viewName, fileName: constants.FILE_NAME.LLMS_CONFIG
         });
         let llmsConfig = {};
@@ -1310,7 +1312,7 @@ const loadLlmsTxt = async (req, res) => {
 const previewLlmsTxt = async (req, res) => {
     const { orgName, viewName } = req.params;
     try {
-        const orgDetails = await adminDao.getOrganization(orgName);
+        const orgDetails = await orgDao.get(orgName);
         const orgID = orgDetails.ORG_ID;
         const { portalName, portalDescription } = req.body;
 
@@ -1336,7 +1338,7 @@ const loadAPIsMd = async (req, res) => {
     const { orgName, viewName } = req.params;
 
     try {
-        const orgDetails = await adminDao.getOrganization(orgName);
+        const orgDetails = await orgDao.get(orgName);
         const orgID = orgDetails.ORG_ID;
 
         if (await isAiDisabledForPortal(orgID, viewName)) {
@@ -1385,7 +1387,7 @@ const loadMCPsMd = async (req, res) => {
     const { orgName, viewName } = req.params;
 
     try {
-        const orgDetails = await adminDao.getOrganization(orgName);
+        const orgDetails = await orgDao.get(orgName);
         const orgID = orgDetails.ORG_ID;
 
         if (await isAiDisabledForPortal(orgID, viewName)) {
@@ -1434,7 +1436,7 @@ const SPEC_FORMAT_DEFAULT = { format: 'json', field: 'swagger', label: 'REST' };
 const loadAPIDefinitionRaw = async (req, res) => {
     const { orgName, apiHandle, viewName, format } = req.params;
     try {
-        const orgDetails = await adminDao.getOrganization(orgName);
+        const orgDetails = await orgDao.get(orgName);
         const orgID = orgDetails.ORG_ID;
 
         if (await isAiDisabledForPortal(orgID, viewName)) {
@@ -1500,21 +1502,21 @@ const loadDocumentMd = async (req, res) => {
     const { orgName, apiHandle, viewName, docType, docName } = req.params;
 
     try {
-        const orgDetails = await adminDao.getOrganization(orgName);
+        const orgDetails = await orgDao.get(orgName);
         const orgID = orgDetails.ORG_ID;
 
         if (await isAiDisabledForPortal(orgID, viewName)) {
             return res.status(404).send('# Not Found\n\nThis resource is not available for agents.');
         }
 
-        const apiID = await apiDao.getAPIId(orgID, apiHandle);
+        const apiID = await apiDao.getId(orgID, apiHandle);
         const docMetaData = await loadAPIMetaData(req, orgID, apiID);
         if (docMetaData.apiInfo?.agentVisibility === 'HIDDEN') {
             return res.status(404).send('# Not Found\n\nThis API is not available for agents.');
         }
         // docName here is without the .md suffix (stripped by the route param)
         const fullDocName = docName + '.md';
-        const docContentResponse = await apiDao.getAPIDocByName(
+        const docContentResponse = await apiFileDao.getDocByName(
             constants.DOC_TYPES.DOC_ID + docType,
             fullDocName,
             orgID,
