--- portals/developer-portal/src/dao/apiMetadataDao.js
@@ -1,1886 +1,0 @@
-/*
- * Copyright (c) 2024, WSO2 LLC. (http://www.wso2.com) All Rights Reserved.
- *
- * WSO2 LLC. licenses this file to you under the Apache License,
- * Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-const { APIMetadata, APILabels } = require('../models/apiMetadata');
-const APISubscriptionPolicy = require('../models/apiSubscriptionPolicy');
-const SubscriptionPolicy = require('../models/subscriptionPolicy');
-const APIContent = require('../models/apiContent');
-const APIImageMetadata = require('../models/apiImage');
-const Labels = require('../models/label');
-const View = require('../models/view');
-const ViewLabels = require('../models/viewLabel');
-const { Sequelize } = require('sequelize');
-const { Op } = require('sequelize');
-const constants = require('../utils/constants');
-const { CustomError } = require('../utils/errors/customErrors');
-const logger = require('../config/logger');
-const fs = require('fs');
-const path = require('path');
-
-const SEARCH_APIS_POSTGRES_SQL = fs.readFileSync(
-    path.join(__dirname, '../../database/queries/search-apis.postgres.sql'),
-    'utf8'
-);
-
-const createAPIMetadata = async (orgID, apiMetadata, t) => {
-
-    const apiInfo = apiMetadata.apiInfo;
-    let owners = {};
-    if (apiInfo.owners) {
-        owners = apiInfo.owners;
-    }
-    try {
-        const apiMetadataResponse = await APIMetadata.create({
-            REFERENCE_ID: apiInfo.referenceID,
-            STATUS: apiInfo.apiStatus,
-            PROVIDER: apiInfo.provider,
-            API_NAME: apiInfo.apiName,
-            API_HANDLE: apiInfo.apiHandle ? apiInfo.apiHandle : `${apiInfo.apiName.toLowerCase().replace(/\s+/g, '')}-v${apiInfo.apiVersion}`,
-            API_DESCRIPTION: apiInfo.apiDescription,
-            API_VERSION: apiInfo.apiVersion,
-            API_TYPE: apiInfo.apiType,
-            VISIBILITY: apiInfo.visibility,
-            VISIBLE_GROUPS: apiInfo.visibleGroups ? apiInfo.visibleGroups.join(' ') : null,
-            AGENT_VISIBILITY: apiMetadata.agentVisibility || apiInfo.agentVisibility || 'VISIBLE',
-            TAGS: apiInfo.tags ? apiInfo.tags.join(' ') : null,
-            TECHNICAL_OWNER: owners.technicalOwner,
-            TECHNICAL_OWNER_EMAIL: owners.technicalOwnerEmail,
-            BUSINESS_OWNER_EMAIL: owners.businessOwnerEmail,
-            BUSINESS_OWNER: owners.businessOwner,
-            SANDBOX_URL: apiMetadata.endPoints.sandboxURL,
-            PRODUCTION_URL: apiMetadata.endPoints.productionURL,
-            METADATA_SEARCH: apiMetadata,
-            GATEWAY_TYPE: apiMetadata.apiInfo.gatewayType || null,
-            ORG_ID: orgID
-        },
-            { transaction: t }
-        );
-        return apiMetadataResponse;
-    } catch (error) {
-        if (error instanceof Sequelize.UniqueConstraintError) {
-            throw error;
-        }
-        throw new Sequelize.DatabaseError(error);
-    }
-};
-
-const createAPILabelMapping = async (orgID, apiID, labels, t) => {
-
-    const labelList = [];
-    const IDList = await getLabelID(orgID, labels, t);
-    try {
-        IDList.forEach(label => {
-            labelList.push({
-                LABEL_ID: label,
-                API_ID: apiID,
-                ORG_ID: orgID
-            });
-        });
-        const labelResponse = await APILabels.bulkCreate(labelList, { transaction: t, ignoreDuplicates: true });
-        return labelResponse;
-    } catch (error) {
-        if (error instanceof Sequelize.UniqueConstraintError) {
-            throw error;
-        }
-        throw new Sequelize.DatabaseError(error);
-    }
-
-}
-
-const createLabels = async (orgID, labels, t) => {
-
-    const labelList = [];
-    try {
-        labels.forEach(label => {
-            labelList.push({
-                NAME: label.name,
-                DISPLAY_NAME: label.displayName,
-                ORG_ID: orgID
-            });
-        })
-        const labelResponse = await Labels.bulkCreate(labelList, { transaction: t });
-        return labelResponse;
-    } catch (error) {
-        if (error instanceof Sequelize.UniqueConstraintError) {
-            throw error;
-        }
-        throw new Sequelize.DatabaseError(error);
-    }
-}
-
-const updateLabel = async (orgID, label, t) => {
-
-    try {
-        let [record, created] = await Labels.findOrCreate({
-            where: {
-                NAME: label.name,
-                ORG_ID: orgID
-            },
-            defaults: {
-                NAME: label.name,
-                DISPLAY_NAME: label.displayName,
-            },
-            transaction: t,
-            returning: true
-        });
-        if (!created) {
-            record = await record.update(label, { transaction: t }); // Update if found
-        }
-        return record;
-    } catch (error) {
-        if (error instanceof Sequelize.UniqueConstraintError) {
-            throw error;
-        }
-        throw new Sequelize.DatabaseError(error);
-    }
-}
-
-const getLabelID = async (orgID, labels, t) => {
-
-    let IDList = [];
-    try {
-        for (const label of labels) {
-            IDList.push(await getLabelIDList(orgID, label, t));
-        };
-        return IDList;
-    } catch (error) {
-        if (error instanceof Sequelize.UniqueConstraintError || error instanceof CustomError) {
-            throw error;
-        }
-        throw new Sequelize.DatabaseError(error);
-    }
-}
-
-const getLabelIDList = async (orgID, label, t) => {
-
-    const labelResponse = await Labels.findOne({
-        where: {
-            NAME: label,
-            ORG_ID: orgID
-        }
-    }, { transaction: t });
-    if (!labelResponse) {
-        throw new CustomError(404, constants.ERROR_CODE[404], "Label not found")
-    }
-    return labelResponse.dataValues.LABEL_ID;
-}
-
-const deleteLabel = async (orgID, labelNames) => {
-
-    try {
-        const labelResponse = await Labels.destroy({
-            where: {
-                NAME: labelNames,
-                ORG_ID: orgID
-            }
-        });
-        return labelResponse;
-    } catch (error) {
-        if (error instanceof Sequelize.UniqueConstraintError) {
-            throw error;
-        }
-        throw new Sequelize.DatabaseError(error);
-    }
-}
-
-const getLabels = async (orgID) => {
-
-    try {
-        const labelResponse = await Labels.findAll({
-            where: {
-                ORG_ID: orgID
-            }
-        });
-        return labelResponse;
-    } catch (error) {
-        if (error instanceof Sequelize.UniqueConstraintError) {
-            throw error;
-        }
-        throw new Sequelize.DatabaseError(error);
-    }
-}
-
-const addView = async (orgID, payload, t) => {
-
-    let displayName = payload.displayName ? payload.displayName : payload.name;
-    try {
-        const viewResponse = await View.create({
-            DISPLAY_NAME: displayName,
-            NAME: payload.name,
-            ORG_ID: orgID
-        }, { transaction: t });
-        return viewResponse;
-    } catch (error) {
-        if (error instanceof Sequelize.UniqueConstraintError) {
-            throw error;
-        }
-        throw new Sequelize.DatabaseError(error);
-    }
-}
-
-const updateView = async (orgID, name, displayName, t) => {
-
-    try {
-        let [record, created] = await View.findOrCreate({
-            where: {
-                NAME: name,
-                ORG_ID: orgID
-            },
-            defaults: {
-                NAME: name,
-                DISPLAY_NAME: displayName,
-            },
-            transaction: t,
-            returning: true
-        });
-        if (!created) {
-            record = await record.update({
-                NAME: name,
-                DISPLAY_NAME: displayName,
-            }, { transaction: t }); // Update if found
-        }
-        return record;
-    } catch (error) {
-        if (error instanceof Sequelize.UniqueConstraintError) {
-            throw error;
-        }
-        throw new Sequelize.DatabaseError(error);
-    }
-}
-
-const deleteView = async (orgID, viewName) => {
-
-    try {
-        const viewResponse = await View.destroy({
-            where: {
-                NAME: viewName,
-                ORG_ID: orgID
-            }
-        });
-        return viewResponse;
-    } catch (error) {
-        if (error instanceof Sequelize.UniqueConstraintError) {
-            throw error;
-        }
-        throw new Sequelize.DatabaseError(error);
-    }
-}
-
-const getView = async (orgID, viewName) => {
-
-    try {
-        const viewResponse = await View.findOne({
-            where: {
-                NAME: viewName,
-                ORG_ID: orgID
-            },
-            include: {
-                model: Labels,
-                attributes: ["NAME"],
-                through: { attributes: [] }
-            },
-        });
-        return viewResponse;
-    } catch (error) {
-        if (error instanceof Sequelize.UniqueConstraintError) {
-            throw error;
-        }
-        throw new Sequelize.DatabaseError(error);
-    }
-}
-
-const getViewID = async (orgID, viewName) => {
-
-    try {
-        const viewResponse = await View.findOne({
-            where: {
-                [Op.or]: [
-                    { DISPLAY_NAME: viewName },
-                    { NAME: viewName }
-                ],
-                ORG_ID: orgID
-            }
-        });
-        if (!viewResponse) {
-            throw new CustomError(404, constants.ERROR_CODE[404], "View not found")
-        }
-        return viewResponse.dataValues.VIEW_ID;
-    } catch (error) {
-        if (error instanceof Sequelize.UniqueConstraintError) {
-            throw error;
-        }
-        throw error;
-    }
-}
-
-const getAllViews = async (orgID) => {
-
-    try {
-        const viewResponse = await View.findAll({
-            where: {
-                ORG_ID: orgID
-            },
-            include: {
-                model: Labels,
-                attributes: ["NAME"],
-                through: {
-                    attributes: []
-                }
-            },
-        });
-        return viewResponse;
-    } catch (error) {
-        if (error instanceof Sequelize.UniqueConstraintError) {
-            throw error;
-        }
-        throw new Sequelize.DatabaseError(error);
-    }
-}
-
-const addViewLabels = async (orgID, viewID, labels, t) => {
-
-    const labelList = [];
-    const IDList = await getLabelID(orgID, labels, t);
-    try {
-        IDList.forEach(label => {
-            labelList.push({
-                LABEL_ID: label,
-                VIEW_ID: viewID,
-                ORG_ID: orgID
-            });
-        });
-        const labelResponse = await ViewLabels.bulkCreate(labelList, { transaction: t });
-        return labelResponse;
-    } catch (error) {
-        if (error instanceof Sequelize.UniqueConstraintError) {
-            throw error;
-        }
-        throw new Sequelize.DatabaseError(error);
-    }
-}
-
-const addLabel = async (orgID, labelID, viewID, t) => {
-
-    try {
-        const labelResponse = await ViewLabels.create({
-            LABEL_ID: labelID,
-            VIEW_ID: viewID,
-            ORG_ID: orgID
-        }, { transaction: t });
-        return labelResponse;
-    } catch (error) {
-        if (error instanceof Sequelize.UniqueConstraintError) {
-            throw error;
-        }
-        throw new Sequelize.DatabaseError(error);
-    }
-}
-
-const replaceViewLabels = async (orgID, viewID, labelNames, t) => {
-    try {
-        await ViewLabels.destroy({ where: { VIEW_ID: viewID, ORG_ID: orgID }, transaction: t });
-        if (labelNames?.length) {
-            await addViewLabels(orgID, viewID, labelNames, t);
-        }
-    } catch (error) {
-        if (error instanceof Sequelize.UniqueConstraintError || error instanceof CustomError) {
-            throw error;
-        }
-        throw new Sequelize.DatabaseError(error);
-    }
-}
-
-const deleteViewLabels = async (orgID, viewID, labels, t) => {
-
-    const IDList = await getLabelID(orgID, labels);
-    let deleteResponse;
-    try {
-        IDList.forEach(async label => {
-            deleteResponse = await ViewLabels.destroy({
-                where: {
-                    LABEL_ID: label,
-                    VIEW_ID: viewID,
-                    ORG_ID: orgID
-                }
-            }, { transaction: t });
-        });
-        return deleteResponse;
-    } catch (error) {
-        if (error instanceof Sequelize.UniqueConstraintError) {
-            throw error;
-        }
-        throw new Sequelize.DatabaseError(error);
-    }
-}
-
-
-const deleteAPILabels = async (orgID, apiID, labels, t) => {
-
-    const IDList = await getLabelID(orgID, labels);
-    let deleteResponse;
-    try {
-        IDList.forEach(async label => {
-            deleteResponse = await APILabels.destroy({
-                where: {
-                    LABEL_ID: label,
-                    API_ID: apiID,
-                    ORG_ID: orgID
-                }
-            }, { transaction: t });
-        });
-        return deleteResponse;
-    } catch (error) {
-        if (error instanceof Sequelize.UniqueConstraintError) {
-            throw error;
-        }
-        throw new Sequelize.DatabaseError(error);
-    }
-}
-
-const toUpper = (v) => (v ? String(v).toUpperCase() : null);
-
-const computeRequestCount = (policy) => {
-  const type = (policy.type || "").toLowerCase();
-
-  if (type === "requestcount") {
-    return policy.requestCount === -1 ? "Unlimited" : String(policy.requestCount);
-  }
-  if (type === "eventcount") {
-    return policy.eventCount === -1 ? "Unlimited" : String(policy.eventCount);
-  }
-  return null;
-};
-
-const buildSubscriptionPolicyRow = (orgID, policy) => {
-  const requestCount = computeRequestCount(policy);
-
-  return {
-    ORG_ID: orgID,
-
-    // Store the APIM policy UUID if provided
-    POLICY_ID: policy.policyId ?? policy.policyID ?? undefined,
-
-    POLICY_NAME: policy.policyName,
-    DISPLAY_NAME: policy.displayName,
-    DESCRIPTION: policy.description,
-    REQUEST_COUNT: requestCount,
-    REF_ID: policy.refId ?? null,
-  };
-};
-
-const createAPISubscriptionPolicy = async (apiSubscriptionPolicies, apiID, t) => {
-  try {
-    const rows = apiSubscriptionPolicies.map((policy) => ({
-      POLICY_ID: policy.policyId ?? policy.policyID,
-      API_ID: apiID,
-    }));
-
-    return await APISubscriptionPolicy.bulkCreate(rows, { transaction: t });
-  } catch (error) {
-    if (error instanceof Sequelize.ValidationError) throw error;
-    throw new Sequelize.DatabaseError(error);
-  }
-};
-
-const putSubscriptionPolicy = async (orgID, policy, t) => {
-  const current = await getSubscriptionPolicyByName(orgID, policy.policyName, t);
-  if (current) {
-    const updated = await updateSubscriptionPolicy(orgID, current.POLICY_ID, policy, t);
-    return { subscriptionPolicyResponse: updated, statusCode: 200 };
-  }
-  const created = await createSubscriptionPolicy(orgID, policy, t);
-  return { subscriptionPolicyResponse: created, statusCode: 201 };
-};
-
-
-const createSubscriptionPolicy = async (orgID, policy, t) => {
-  try {
-    const row = buildSubscriptionPolicyRow(orgID, policy);
-
-    return await SubscriptionPolicy.create(row, { transaction: t });
-  } catch (error) {
-    if (error instanceof Sequelize.UniqueConstraintError || error instanceof Sequelize.ValidationError) {
-      throw error;
-    }
-    throw new Sequelize.DatabaseError(error);
-  }
-};
-
-/**
- * Bulk create subscription policies
- * @param {} orgID 
- * @param {*} policies 
- * @param {*} t 
- * @returns 
- */
-const bulkCreateSubscriptionPolicies = async (orgID, policies, t) => {
-  try {
-    const rows = policies.map((policy) => buildSubscriptionPolicyRow(orgID, policy));
-
-    return await SubscriptionPolicy.bulkCreate(rows, { transaction: t });
-  } catch (error) {
-    if (error instanceof Sequelize.UniqueConstraintError || error instanceof Sequelize.ValidationError) {
-      throw error;
-    }
-    throw new Sequelize.DatabaseError(error);
-  }
-};
-
-const updateSubscriptionPolicy = async (orgID, policyID, policy, t) => {
-  try {
-    const row = buildSubscriptionPolicyRow(orgID, policy);
-
-    // Don’t update primary keys
-    delete row.ORG_ID;
-    delete row.POLICY_ID;
-    if (!Object.prototype.hasOwnProperty.call(policy, 'refId')) {
-      delete row.REF_ID;
-    }
-
-    const [_, updatedRows] = await SubscriptionPolicy.update(row, {
-      where: { POLICY_ID: policyID, ORG_ID: orgID },
-      returning: true,
-      transaction: t
-    });
-
-    return updatedRows[0];
-  } catch (error) {
-    if (error instanceof Sequelize.UniqueConstraintError || error instanceof Sequelize.ValidationError) {
-      throw error;
-    }
-    throw new Sequelize.DatabaseError(error);
-  }
-};
-
-
-const deleteSubscriptionPolicy = async (orgID, policyName, t) => {
-
-    try {
-        const subscriptionPolicyResponse = await SubscriptionPolicy.destroy({
-            where: {
-                POLICY_NAME: policyName,
-                ORG_ID: orgID
-            },
-            transaction: t
-        });
-        return subscriptionPolicyResponse;
-    } catch (error) {
-        if (error instanceof Sequelize.ValidationError) {
-            throw error;
-        }
-        throw new Sequelize.DatabaseError(error);
-    }
-}
-
-
-const deleteSubscriptionPolicyById = async (orgID, policyID, t) => {
-
-    try {
-        const subscriptionPolicyResponse = await SubscriptionPolicy.destroy({
-            where: {
-                POLICY_ID: policyID,
-                ORG_ID: orgID
-            },
-            transaction: t
-        });
-        return subscriptionPolicyResponse;
-    } catch (error) {
-        if (error instanceof Sequelize.ValidationError) {
-            throw error;
-        }
-        throw new Sequelize.DatabaseError(error);
-    }
-}
-
-
-const getSubscriptionPolicyByName = async (orgID, policyName, t) => {
-
-    try {
-        const subscriptionPolicyResponse = await SubscriptionPolicy.findOne({
-            where: {
-                POLICY_NAME: policyName,
-                ORG_ID: orgID
-            },
-            transaction: t
-        });
-        return subscriptionPolicyResponse;
-    } catch (error) {
-        if (error instanceof Sequelize.ValidationError) {
-            throw error;
-        }
-        throw new Sequelize.DatabaseError(error);
-    }
-};
-
-
-const storeAPIImageMetadata = async (apiImages, apiID, t) => {
-
-    let apiImagesList = [];
-    try {
-        for (var propertyKey in apiImages) {
-            apiImagesList.push({
-                IMAGE_TAG: propertyKey,
-                IMAGE_NAME: apiImages[propertyKey],
-                API_ID: apiID
-            })
-        }
-        const apiImagesResponse = await APIImageMetadata.bulkCreate(apiImagesList, { transaction: t });
-        return apiImagesResponse;
-    } catch (error) {
-        if (error instanceof Sequelize.UniqueConstraintError) {
-            throw error;
-        }
-        throw new Sequelize.DatabaseError(error);
-    }
-}
-
-const storeAPIFile = async (apiFile, fileName, apiID, type, t) => {
-
-    try {
-        const apiFileResponse = await APIContent.create({
-            API_FILE: apiFile,
-            FILE_NAME: fileName,
-            API_ID: apiID,
-            TYPE: type
-        }, { transaction: t }
-        );
-        return apiFileResponse;
-    } catch (error) {
-        if (error instanceof Sequelize.UniqueConstraintError) {
-            throw error;
-        }
-        throw new Sequelize.DatabaseError(error);
-    }
-}
-
-const storeAPIFiles = async (files, apiID, t) => {
-
-    let apiContent = []
-    try {
-        files.forEach(file => {
-            apiContent.push({
-                API_FILE: file.content,
-                FILE_NAME: file.fileName,
-                TYPE: file.type,
-                API_ID: apiID
-            })
-        });
-        const apiContentResponse = await APIContent.bulkCreate(apiContent, { transaction: t });
-        return apiContentResponse;
-    } catch (error) {
-        if (error instanceof Sequelize.UniqueConstraintError) {
-            throw error;
-        }
-        throw new Sequelize.DatabaseError(error);
-    }
-}
-
-const updateOrCreateAPIFiles = async (files, apiID, orgID, t) => {
-
-    let filesToCreate = []
-    try {
-        for (const file of files) {
-            const apiFileResponse = await getAPIFile(file.fileName, file.type, orgID, apiID, t);
-            if (apiFileResponse == null || apiFileResponse == undefined) {
-                filesToCreate.push({
-                    API_FILE: file.content,
-                    FILE_NAME: file.fileName,
-                    API_ID: apiID,
-                    TYPE: file.type,
-                })
-            } else {
-                const updateResponse = await APIContent.update(
-                    {
-                        API_FILE: file.content,
-                    },
-                    {
-                        where: {
-                            API_ID: apiID,
-                            FILE_NAME: apiFileResponse.FILE_NAME,
-                            TYPE: file.type,
-                        },
-                        include: [
-                            {
-                                model: APIMetadata,
-                                where: {
-                                    ORG_ID: orgID
-                                }
-                            }
-                        ]
-                    }
-                );
-                if (!updateResponse) {
-                    throw new Sequelize.DatabaseError('Error while updating API files');
-                }
-            }
-        };
-        if (filesToCreate.length > 0) {
-            await APIContent.bulkCreate(filesToCreate, { transaction: t });
-        }
-    } catch (error) {
-        if (error instanceof Sequelize.UniqueConstraintError) {
-            throw error;
-        }
-        throw new Sequelize.DatabaseError(error);
-    }
-}
-
-const getAPIFile = async (fileName, type, orgID, apiID, t) => {
-
-    try {
-        const apiFileResponse = await APIContent.findOne({
-            where: {
-                FILE_NAME: fileName,
-                API_ID: apiID,
-                TYPE: type
-            },
-            include: [
-                {
-                    model: APIMetadata,
-                    where: {
-                        ORG_ID: orgID
-                    }
-                }
-            ],
-            transaction: t
-        });
-        return apiFileResponse;
-    } catch (error) {
-        if (error instanceof Sequelize.UniqueConstraintError) {
-            throw error;
-        }
-        throw new Sequelize.DatabaseError(error);
-    }
-}
-
-const getAPIFileByType = async (type, orgID, apiID, t) => {
-    try {
-        const apiFileResponse = await APIContent.findOne({
-            where: {
-                API_ID: apiID,
-                TYPE: type
-            },
-            include: [
-                {
-                    model: APIMetadata,
-                    where: {
-                        ORG_ID: orgID
-                    }
-                }
-            ],
-            transaction: t
-        });
-        return apiFileResponse;
-    } catch (error) {
-        if (error instanceof Sequelize.UniqueConstraintError) {
-            throw error;
-        }
-        throw new Sequelize.DatabaseError(error);
-    }
-}
-
-const upsertAPIFileByType = async (apiFile, fileName, apiID, orgID, type, t) => {
-    try {
-        const apiFileResponse = await getAPIFileByType(type, orgID, apiID, t);
-        let fileUpdateResponse;
-        if (apiFileResponse == null || apiFileResponse == undefined) {
-            fileUpdateResponse = await APIContent.create({
-                API_FILE: apiFile,
-                FILE_NAME: fileName,
-                API_ID: apiID,
-                TYPE: type
-            }, { transaction: t });
-        } else {
-            fileUpdateResponse = await APIContent.update({
-                API_FILE: apiFile,
-                FILE_NAME: fileName
-            },
-                {
-                    where: {
-                        API_ID: apiID,
-                        TYPE: type
-                    },
-                    include: [
-                        {
-                            model: APIMetadata,
-                            where: {
-                                ORG_ID: orgID
-                            }
-                        }
-                    ],
-                    transaction: t
-                }
-            );
-        }
-        return fileUpdateResponse;
-    } catch (error) {
-        if (error instanceof Sequelize.UniqueConstraintError) {
-            throw error;
-        }
-        throw new Sequelize.DatabaseError(error);
-    }
-}
-
-const getAPIDoc = async (type, orgID, apiID, t) => {
-
-    try {
-        const apiFileResponse = await APIContent.findOne({
-            where: {
-                API_ID: apiID,
-                TYPE: type
-            },
-            include: [
-                {
-                    model: APIMetadata,
-                    where: {
-                        ORG_ID: orgID
-                    }
-                }
-            ],
-            transaction: t
-        });
-        return apiFileResponse;
-    } catch (error) {
-        if (error instanceof Sequelize.UniqueConstraintError) {
-            throw error;
-        }
-        throw new Sequelize.DatabaseError(error);
-    }
-}
-
-const getAPIDocByName = async (type, name, orgID, apiID, t) => {
-
-    try {
-        const apiFileResponse = await APIContent.findOne({
-            where: {
-                API_ID: apiID,
-                TYPE: type,
-                FILE_NAME: name
-            },
-            include: [
-                {
-                    model: APIMetadata,
-                    where: {
-                        ORG_ID: orgID
-                    }
-                }
-            ], transaction: t
-        });
-        return apiFileResponse;
-    } catch (error) {
-        if (error instanceof Sequelize.UniqueConstraintError) {
-            throw error;
-        }
-        throw new Sequelize.DatabaseError(error);
-    }
-}
-
-const getAPIDocTypes = async (orgID, apiID) => {
-    const isPostgres = APIContent.sequelize.getDialect() === 'postgres';
-    const fileNamesExpr = isPostgres
-        ? [Sequelize.fn("ARRAY_AGG", Sequelize.col("DP_API_CONTENT.FILE_NAME")), "FILE_NAMES"]
-        : [Sequelize.fn("GROUP_CONCAT", Sequelize.col("DP_API_CONTENT.FILE_NAME"), "|||"), "FILE_NAMES"];
-
-    try {
-        const rows = await APIContent.findAll({
-            attributes: ["TYPE", fileNamesExpr],
-            where: {
-                API_ID: apiID,
-                TYPE: {
-                    [Op.or]: [
-                        { [Op.like]: "DOC_%" },
-                        { [Op.like]: constants.DOC_TYPES.API_DEFINITION }
-                    ]
-                },
-            },
-            group: ["DP_API_CONTENT.TYPE"],
-            include: [
-                {
-                    model: APIMetadata,
-                    required: true,
-                    attributes: [],
-                    where: {
-                        ORG_ID: orgID
-                    }
-                }
-            ]
-        });
-
-        if (!isPostgres) {
-            for (const row of rows) {
-                const raw = row.dataValues.FILE_NAMES;
-                row.dataValues.FILE_NAMES = raw ? raw.split("|||") : [];
-            }
-        }
-
-        return rows;
-    } catch (error) {
-        if (error instanceof Sequelize.UniqueConstraintError) {
-            throw error;
-        }
-        throw new Sequelize.DatabaseError(error);
-    }
-}
-
-const getAPIDocs = async (orgID, apiID) => {
-    const isPostgres = APIContent.sequelize.getDialect() === 'postgres';
-    const include = [{
-        model: APIMetadata,
-        required: true,
-        attributes: [],
-        where: { ORG_ID: orgID }
-    }];
-    const where = {
-        API_ID: apiID,
-        [Op.or]: [
-            { TYPE: { [Op.like]: "DOC_%" } },
-            { FILE_NAME: { [Op.like]: "LINK_%" } }
-        ]
-    };
-
-    try {
-        if (isPostgres) {
-            return await APIContent.findAll({
-                attributes: [
-                    "TYPE",
-                    [Sequelize.fn("ARRAY_AGG", Sequelize.col("DP_API_CONTENT.FILE_NAME")), "FILE_NAMES"],
-                    [Sequelize.fn("ARRAY_AGG", Sequelize.col("DP_API_CONTENT.API_FILE")), "API_FILES"]
-                ],
-                where,
-                group: ["DP_API_CONTENT.TYPE"],
-                include
-            });
-        }
-
-        const rows = await APIContent.findAll({ attributes: ["TYPE", "FILE_NAME", "API_FILE"], where, include });
-        const typeMap = new Map();
-        for (const row of rows) {
-            const { TYPE, FILE_NAME, API_FILE } = row.dataValues;
-            if (!typeMap.has(TYPE)) typeMap.set(TYPE, { TYPE, FILE_NAMES: [], API_FILES: [] });
-            typeMap.get(TYPE).FILE_NAMES.push(FILE_NAME);
-            typeMap.get(TYPE).API_FILES.push(API_FILE);
-        }
-        return Array.from(typeMap.values()).map(g => ({ dataValues: g }));
-    } catch (error) {
-        if (error instanceof Sequelize.UniqueConstraintError) {
-            throw error;
-        }
-        throw new Sequelize.DatabaseError(error);
-    }
-}
-
-
-const getAPIDocLinks = async (orgID, apiID) => {
-    const isPostgres = APIContent.sequelize.getDialect() === 'postgres';
-    const include = [{
-        model: APIMetadata,
-        required: true,
-        attributes: [],
-        where: { ORG_ID: orgID }
-    }];
-    const where = {
-        API_ID: apiID,
-        FILE_NAME: { [Op.like]: "LINK_%" }
-    };
-
-    try {
-        if (isPostgres) {
-            return await APIContent.findAll({
-                attributes: [
-                    "TYPE",
-                    [Sequelize.fn("ARRAY_AGG", Sequelize.col("DP_API_CONTENT.FILE_NAME")), "FILE_NAMES"],
-                    [Sequelize.fn("ARRAY_AGG", Sequelize.col("DP_API_CONTENT.API_FILE")), "API_FILES"]
-                ],
-                where,
-                group: ["DP_API_CONTENT.TYPE"],
-                include
-            });
-        }
-
-        const rows = await APIContent.findAll({ attributes: ["TYPE", "FILE_NAME", "API_FILE"], where, include });
-        const typeMap = new Map();
-        for (const row of rows) {
-            const { TYPE, FILE_NAME, API_FILE } = row.dataValues;
-            if (!typeMap.has(TYPE)) typeMap.set(TYPE, { TYPE, FILE_NAMES: [], API_FILES: [] });
-            typeMap.get(TYPE).FILE_NAMES.push(FILE_NAME);
-            typeMap.get(TYPE).API_FILES.push(API_FILE);
-        }
-        return Array.from(typeMap.values()).map(g => ({ dataValues: g }));
-    } catch (error) {
-        if (error instanceof Sequelize.UniqueConstraintError) {
-            throw error;
-        }
-        throw new Sequelize.DatabaseError(error);
-    }
-}
-
-const getAPISpecs = async (orgID, apiIDs) => {
-    try {
-        const apiSpecsResponse = await APIContent.findAll({
-            attributes: [
-                'API_ID',
-                'FILE_NAME',
-                'API_FILE'
-            ],
-            where: {
-                API_ID: {
-                    [Op.in]: apiIDs
-                },
-                TYPE: constants.DOC_TYPES.API_DEFINITION
-            },
-            include: [
-                {
-                    model: APIMetadata,
-                    required: true,
-                    attributes: ['API_NAME', 'API_VERSION', 'API_HANDLE'],
-                    where: {
-                        ORG_ID: orgID
-                    }
-                }
-            ]
-        });
-
-        return apiSpecsResponse.map(spec => {
-
-            return {
-                apiID: spec.API_ID,
-                fileName: spec.FILE_NAME,
-                apiSpec: spec.API_FILE ? spec.API_FILE.toString('utf8') : null
-            };
-        }).filter(spec => spec !== null);
-    } catch (error) {
-        logger.error('Error fetching API specifications', {
-            error: error.message,
-            stack: error.stack,
-            operation: 'fetchAPISpecifications'
-        });
-        if (error instanceof Sequelize.UniqueConstraintError) {
-            throw error;
-        }
-        throw new Sequelize.DatabaseError(error);
-    }
-}
-
-const getAPIMetadataByCondition = async (condition, t) => {
-    try {
-        if (condition.TAGS) {
-            const tagsArray = condition.TAGS.split(",").map(tag => tag.trim());
-            condition.TAGS = {
-                [Op.or]: tagsArray.map(tag => ({
-                    [Op.and]: {
-                        [Sequelize.Op.or]: [
-                            {
-                                [Sequelize.Op.like]: `% ${tag} %`
-                            },
-                            {
-                                [Sequelize.Op.like]: `% ${tag}`
-                            },
-                            {
-                                [Sequelize.Op.like]: `${tag} %`
-                            },
-                            {
-                                [Sequelize.Op.eq]: `${tag}`
-                            }
-                        ]
-                    }
-                }))
-            };
-        }
-        const apiMetadataResponse = await APIMetadata.findAll({
-            include: [{
-                model: APIImageMetadata,
-                required: false
-            }, {
-                model: SubscriptionPolicy,
-                through: { attributes: [] },
-                required: false
-            }
-            ],
-            where: condition,
-            transaction: t
-        });
-        return apiMetadataResponse;
-    } catch (error) {
-        if (error instanceof Sequelize.UniqueConstraintError) {
-            throw error;
-        }
-        throw new Sequelize.DatabaseError(error);
-    }
-}
-
-const getAPIMetadata = async (orgID, apiID, t) => {
-
-    try {
-        const apiMetadataResponse = await APIMetadata.findAll({
-            include: [{
-                model: APIImageMetadata,
-                where: {
-                    API_ID: apiID
-                },
-                required: false
-            }, {
-                model: SubscriptionPolicy,
-                through: { attributes: [] },
-                required: false
-            },
-            {
-                model: Labels,
-                attributes: ["NAME"],
-                through: { attributes: [] }
-            }
-            ],
-            where: {
-                ORG_ID: orgID,
-                API_ID: apiID,
-                STATUS: constants.API_STATUS.PUBLISHED
-            },
-            transaction: t
-        });
-        return apiMetadataResponse;
-    } catch (error) {
-        if (error instanceof Sequelize.UniqueConstraintError) {
-            throw error;
-        }
-        throw new Sequelize.DatabaseError(error);
-    }
-};
-
-const getAllAPIMetadata = async (orgID, groups, viewName, t) => {
-
-    const viewID = await getViewID(orgID, viewName);
-    let apiList = [];
-    for (const group of groups) {
-        try {
-            const apiMetadataResponse = await APIMetadata.findAll({
-                where: {
-                    ORG_ID: orgID,
-                    VISIBLE_GROUPS: {
-                        [Op.like]: `%${group}%`
-                    },
-                    STATUS: constants.API_STATUS.PUBLISHED
-                },
-                include: [{
-                    model: APIImageMetadata,
-                    required: false
-                }, {
-                    model: SubscriptionPolicy,
-                    through: { attributes: [] },
-                    required: false
-                },
-                {
-                    model: Labels,
-                    attributes: ["NAME"],
-                    required: true,
-                    through: { attributes: [] },
-                    where: {
-                        LABEL_ID: {
-                            [Op.in]: Sequelize.literal(`(SELECT "LABEL_ID" FROM "DP_VIEW_LABELS" WHERE "VIEW_ID" = '${viewID}')`)
-                        }
-                    }
-                }
-                ],
-                transaction: t
-            });
-            if (apiMetadataResponse) {
-                apiList.push(...apiMetadataResponse);
-            }
-        } catch (error) {
-            {
-                if (error instanceof Sequelize.UniqueConstraintError) {
-                    throw error;
-                }
-                throw new Sequelize.DatabaseError(error);
-            }
-        }
-    }
-    // add all public apis
-    try {
-        const publicAPIS = await APIMetadata.findAll({
-            where: {
-                ORG_ID: orgID,
-                STATUS: constants.API_STATUS.PUBLISHED
-            },
-            include: [{
-                model: APIImageMetadata,
-                required: false
-            }, {
-                model: SubscriptionPolicy,
-                through: { attributes: [] },
-                required: false
-            },
-            {
-                model: Labels,
-                attributes: ["NAME"],
-                required: true,
-                through: { attributes: [] },
-                where: {
-                    LABEL_ID: {
-                        [Op.in]: Sequelize.literal(`(SELECT "LABEL_ID" FROM "DP_VIEW_LABELS" WHERE "VIEW_ID" = '${viewID}')`)
-                    }
-                }
-            }
-            ],
-            transaction: t
-        });
-        apiList.push(...publicAPIS);
-    } catch (error) {
-        {
-            if (error instanceof Sequelize.UniqueConstraintError) {
-                throw error;
-            }
-            throw new Sequelize.DatabaseError(error);
-        }
-    }
-    return apiList;
-};
-
-const getAllAPIMetadataFromAllViews = async (orgID, groups, t) => {
-
-    let apiList = [];
-    for (const group of groups) {
-        try {
-            const apiMetadataResponse = await APIMetadata.findAll({
-                where: {
-                    ORG_ID: orgID,
-                    VISIBLE_GROUPS: {
-                        [Op.like]: `%${group}%`
-                    },
-                    STATUS: constants.API_STATUS.PUBLISHED
-                },
-                include: [{
-                    model: APIImageMetadata,
-                    required: false
-                }, {
-                    model: SubscriptionPolicy,
-                    through: { attributes: [] },
-                    required: false
-                },
-                {
-                    model: Labels,
-                    attributes: ["NAME"],
-                    required: false,
-                    through: { attributes: [] }
-                }
-                ],
-                transaction: t
-            });
-            if (apiMetadataResponse) {
-                apiList.push(...apiMetadataResponse);
-            }
-        } catch (error) {
-            {
-                if (error instanceof Sequelize.UniqueConstraintError) {
-                    throw error;
-                }
-                throw new Sequelize.DatabaseError(error);
-            }
-        }
-    }
-    // add all public apis
-    try {
-        const publicAPIS = await APIMetadata.findAll({
-            where: {
-                ORG_ID: orgID,
-                STATUS: constants.API_STATUS.PUBLISHED
-            },
-            include: [{
-                model: APIImageMetadata,
-                required: false
-            }, {
-                model: SubscriptionPolicy,
-                through: { attributes: [] },
-                required: false
-            },
-            {
-                model: Labels,
-                attributes: ["NAME"],
-                required: true,
-                through: { attributes: [] }
-            }
-            ],
-            transaction: t
-        });
-        apiList.push(...publicAPIS);
-    } catch (error) {
-        {
-            if (error instanceof Sequelize.UniqueConstraintError) {
-                throw error;
-            }
-            throw new Sequelize.DatabaseError(error);
-        }
-    }
-    return apiList;
-};
-
-const searchAPIMetadataFallback = async (orgID, searchTerm, viewName, t) => {
-    const pattern = `%${searchTerm}%`;
-    const viewID = await getViewID(orgID, viewName);
-    return APIMetadata.findAll({
-        where: {
-            ORG_ID: orgID,
-            STATUS: constants.API_STATUS.PUBLISHED,
-            [Op.or]: [
-                Sequelize.where(
-                    Sequelize.cast(Sequelize.col('DP_API_METADATA.METADATA_SEARCH'), 'TEXT'),
-                    { [Op.like]: pattern }
-                ),
-                Sequelize.where(
-                    Sequelize.col('DP_API_METADATA.TAGS'),
-                    { [Op.like]: pattern }
-                ),
-            ],
-        },
-        include: [
-            { model: APIImageMetadata, required: false },
-            { model: SubscriptionPolicy, through: { attributes: [] }, required: false },
-            {
-                model: Labels,
-                attributes: ['NAME'],
-                required: true,
-                through: { attributes: [] },
-                where: {
-                    LABEL_ID: {
-                        [Op.in]: Sequelize.literal(`(SELECT "LABEL_ID" FROM "DP_VIEW_LABELS" WHERE "VIEW_ID" = '${viewID}')`)
-                    }
-                }
-            },
-        ],
-        transaction: t,
-    });
-};
-
-const searchAPIMetadata = async (orgID, groups, searchTerm, viewName, t) => {
-    if (APIMetadata.sequelize.getDialect() !== 'postgres') {
-        return searchAPIMetadataFallback(orgID, searchTerm, viewName, t);
-    }
-    try {
-        const results = await APIMetadata.sequelize.query(SEARCH_APIS_POSTGRES_SQL, {
-            replacements: { searchTerm, orgID },
-            type: Sequelize.QueryTypes.SELECT,
-        });
-        return results;
-    } catch (error) {
-        if (error instanceof Sequelize.UniqueConstraintError) {
-            throw error;
-        }
-        throw new Sequelize.DatabaseError(error);
-    }
-};
-
-const deleteAPIMetadata = async (orgID, apiID, t) => {
-
-    try {
-        const apiMetadataResponse = await APIMetadata.destroy({
-            where: {
-                API_ID: apiID,
-                ORG_ID: orgID
-            },
-            transaction: t
-        });
-        return apiMetadataResponse;
-    } catch (error) {
-        if (error instanceof Sequelize.UniqueConstraintError) {
-            throw error;
-        }
-        throw new Sequelize.DatabaseError(error);
-    }
-}
-
-const updateAPIMetadata = async (orgID, apiID, apiMetadata, t) => {
-
-    const apiInfo = apiMetadata.apiInfo;
-    let owners = {};
-    if (apiInfo.owners) {
-        owners = apiInfo.owners;
-    }
-    try {
-        let updateCount, apiMetadataResponse;
-        [updateCount, apiMetadataResponse] = await APIMetadata.update({
-            REFERENCE_ID: apiInfo.referenceID,
-            STATUS: apiInfo.apiStatus,
-            PROVIDER: apiInfo.provider,
-            API_NAME: apiInfo.apiName,
-            API_HANDLE: apiInfo.apiHandle ? apiInfo.apiHandle : `${apiInfo.apiName.toLowerCase().replace(/\s+/g, '')}-v${apiInfo.apiVersion}`,
-            API_DESCRIPTION: apiInfo.apiDescription,
-            API_VERSION: apiInfo.apiVersion,
-            API_TYPE: apiInfo.apiType,
-            TAGS: apiInfo.tags ? apiInfo.tags.join(' ') : null,
-            VISIBILITY: apiInfo.visibility,
-            VISIBLE_GROUPS: apiInfo.visibleGroups ? apiInfo.visibleGroups.join(' ') : null,
-            AGENT_VISIBILITY: apiMetadata.agentVisibility || apiInfo.agentVisibility || 'VISIBLE',
-            TECHNICAL_OWNER: owners.technicalOwner,
-            TECHNICAL_OWNER_EMAIL: owners.technicalOwnerEmail,
-            BUSINESS_OWNER_EMAIL: owners.businessOwnerEmail,
-            BUSINESS_OWNER: owners.businessOwner,
-            SANDBOX_URL: apiMetadata.endPoints.sandboxURL,
-            PRODUCTION_URL: apiMetadata.endPoints.productionURL,
-            METADATA_SEARCH: apiMetadata,
-            GATEWAY_TYPE: apiMetadata.apiInfo.gatewayType || null,
-        }, {
-            where: {
-                API_ID: apiID,
-                ORG_ID: orgID,
-            },
-            returning: true,
-            transaction: t
-        });
-        return [updateCount, apiMetadataResponse];
-    } catch (error) {
-        if (error instanceof Sequelize.UniqueConstraintError) {
-            throw error;
-        }
-        throw new Sequelize.DatabaseError(error);
-    }
-}
-
-async function updateAPISubscriptionPolicy(subscriptionPolicies, apiID, t) {
-
-    let policiesToCreate = [];
-    try {
-        for (const policy of subscriptionPolicies) {
-            policiesToCreate.push({
-                POLICY_ID: policy.policyId ?? policy.policyID,
-                API_ID: apiID,
-            })
-        }
-        if (policiesToCreate.length > 0) {
-            await APISubscriptionPolicy.destroy({
-                where: {
-                    API_ID: apiID
-                },
-                transaction: t
-            });
-            return await APISubscriptionPolicy.bulkCreate(policiesToCreate, { transaction: t });
-        } else {
-            return policiesToCreate;
-        }
-    } catch (error) {
-        if (error instanceof Sequelize.UniqueConstraintError) {
-            throw error;
-        }
-        throw new Sequelize.DatabaseError(error);
-    }
-}
-
-const getSubscriptionPolicy = async (policyID, orgID, t) => {
-    try {
-        const subscriptionPolicyResponse = await SubscriptionPolicy.findOne({
-            where: {
-                ORG_ID: orgID,
-                POLICY_ID: policyID
-            },
-            transaction: t
-        });
-        return subscriptionPolicyResponse;
-    } catch (error) {
-        if (error instanceof Sequelize.EmptyResultError) {
-            throw error;
-        }
-        throw new Sequelize.DatabaseError(error);
-    }
-}
-
-const getSubscriptionPolicies = async (apiID, t) => {
-
-    try {
-        const subscriptionPolicyResponse = await SubscriptionPolicy.findAll({
-            include: [
-                {
-                    model: APIMetadata,
-                    where: { API_ID: apiID },
-                    through: { attributes: [] }
-                }
-            ],
-            transaction: t
-        });
-        return subscriptionPolicyResponse;
-    } catch (error) {
-        if (error instanceof Sequelize.UniqueConstraintError) {
-            throw error;
-        }
-        throw new Sequelize.DatabaseError(error);
-    }
-}
-
-async function getAllSubscriptionPolicies(orgID, t) {
-    try {
-
-        const subscriptionPoliciesResponse = await SubscriptionPolicy.findAll({
-            where: {
-                ORG_ID: orgID
-            },
-            transaction: t
-        });
-        return subscriptionPoliciesResponse;
-    } catch (error) {
-        if (error instanceof Sequelize.UniqueConstraintError) {
-            throw error;
-        }
-        throw new Sequelize.DatabaseError(error);
-    }
-}
-const updateAPIImageMetadata = async (apiImages, orgID, apiID, t) => {
-
-    let imageCreateList = [];
-    try {
-        for (const propertyKey in apiImages) {
-            let apiImageResponse = await getImageMetadata(propertyKey, apiImages[propertyKey], orgID, apiID, t);
-            if (apiImageResponse == null || apiImageResponse == undefined) {
-                imageCreateList.push({
-                    IMAGE_NAME: apiImages[propertyKey],
-                    API_ID: apiID,
-                    IMAGE_TAG: propertyKey
-                })
-            } else {
-                const apiImageDataUpdate = await APIImageMetadata.update({
-                    IMAGE_NAME: apiImages[propertyKey],
-                    IMAGE_TAG: propertyKey
-                }, {
-                    where: {
-                        [Op.or]: [
-                            { IMAGE_TAG: apiImageResponse.IMAGE_TAG },
-                            { IMAGE_NAME: apiImageResponse.IMAGE_NAME }
-                        ],
-                        API_ID: apiID
-                    },
-                    include: [
-                        {
-                            model: APIMetadata,
-                            where: {
-                                ORG_ID: orgID
-                            }
-                        }
-                    ],
-                    transaction: t
-                });
-                if (!apiImageDataUpdate) {
-                    throw new Sequelize.EmptyResultError("Error updating API Image Metadata");
-                }
-            }
-        }
-        if (imageCreateList.length > 0) {
-            await APIImageMetadata.bulkCreate(imageCreateList, { transaction: t });
-        }
-    } catch (error) {
-        if (error instanceof Sequelize.UniqueConstraintError) {
-            throw error;
-        }
-        throw new Sequelize.DatabaseError(error);
-    }
-}
-
-const getImageMetadata = async (imageTag, imageName, orgID, apiID, t) => {
-
-    try {
-        const apiImageData = await APIImageMetadata.findOne({
-            where: {
-                [Op.or]: [
-                    { IMAGE_TAG: imageTag },
-                    { IMAGE_NAME: imageName }
-                ],
-                API_ID: apiID
-            },
-            include: [
-                {
-                    model: APIMetadata,
-                    where: {
-                        ORG_ID: orgID
-                    }
-                }
-            ],
-            transaction: t
-        });
-        return apiImageData;
-    } catch (error) {
-        if (error instanceof Sequelize.UniqueConstraintError) {
-            throw error;
-        }
-        throw new Sequelize.DatabaseError(error);
-    }
-}
-
-const getImage = async (imageTag, apiID, t) => {
-    try {
-        const apiImageData = await APIImageMetadata.findOne({
-            where: {
-                IMAGE_TAG: imageTag,
-                API_ID: apiID
-            },
-            transaction: t
-        });
-        return apiImageData;
-    } catch (error) {
-        if (error instanceof Sequelize.UniqueConstraintError) {
-            throw error;
-        }
-        throw new Sequelize.DatabaseError(error);
-    }
-}
-
-const deleteImage = async (imageTag, apiID, t) => {
-    try {
-        const apiImageData = await APIImageMetadata.destroy({
-            where: {
-                IMAGE_TAG: imageTag,
-                API_ID: apiID
-            },
-            transaction: t
-        });
-        return apiImageData;
-    } catch (error) {
-        if (error instanceof Sequelize.UniqueConstraintError) {
-            throw error;
-        }
-        throw new Sequelize.DatabaseError(error);
-    }
-}
-
-const updateAPIFile = async (apiFile, fileName, apiID, orgID, type, t) => {
-
-    try {
-        const apiFileResponse = await getAPIFile(fileName, type, orgID, apiID, t);
-        let fileUpdateResponse;
-        if (apiFileResponse == null || apiFileResponse == undefined) {
-            fileUpdateResponse = await APIContent.create({
-                API_FILE: apiFile,
-                FILE_NAME: fileName,
-                API_ID: apiID,
-                TYPE: type
-            }, { transaction: t });
-        } else {
-            fileUpdateResponse = await APIContent.update({
-                API_FILE: apiFile,
-                FILE_NAME: fileName
-            },
-                {
-                    where: {
-                        API_ID: apiID,
-                        FILE_NAME: fileName,
-                        TYPE: type
-                    },
-                    include: [
-                        {
-                            model: APIMetadata,
-                            where: {
-                                ORG_ID: orgID
-                            }
-                        }
-                    ]
-                },
-                { transaction: t }
-            );
-        }
-        return fileUpdateResponse;
-    } catch (error) {
-        if (error instanceof Sequelize.UniqueConstraintError) {
-            throw error;
-        }
-        throw new Sequelize.DatabaseError(error);
-    }
-}
-
-const deleteAPIFile = async (fileName, type, orgID, apiID, t) => {
-
-    try {
-        const contentsToDelete = await APIContent.findAll({
-            where: {
-                FILE_NAME: fileName,
-                API_ID: apiID,
-                TYPE: { [Op.like]: `%${type}%` }
-            },
-            include: [
-                {
-                    model: APIMetadata,
-                    where: {
-                        ORG_ID: orgID
-                    }
-                }
-            ],
-            transaction: t
-        });
-        let apiFileResponse;
-        for (const content of contentsToDelete) {
-            apiFileResponse = await APIContent.destroy({
-                where: {
-                    FILE_NAME: content.dataValues.FILE_NAME
-
-                }
-            });
-        }
-        return apiFileResponse;
-    } catch (error) {
-        if (error instanceof Sequelize.UniqueConstraintError) {
-            throw error;
-        }
-        throw new Sequelize.DatabaseError(error);
-    }
-}
-
-const deleteAllAPIFiles = async (type, orgID, apiID, t) => {
-
-    try {
-        const contentsToDelete = await APIContent.findAll({
-            where: {
-                API_ID: apiID,
-                TYPE: {
-                    [Op.like]: `%${type}%`
-                }
-            },
-            include: [
-                {
-                    model: APIMetadata,
-                    where: {
-                        ORG_ID: orgID
-                    }
-                }
-            ],
-            transaction: t
-        });
-        let apiFileResponse;
-        for (const content of contentsToDelete) {
-            apiFileResponse = await APIContent.destroy({
-                where: {
-                    FILE_NAME: content.dataValues.FILE_NAME
-                }
-            });
-        }
-        return apiFileResponse;
-    } catch (error) {
-        if (error instanceof Sequelize.UniqueConstraintError) {
-            throw error;
-        }
-        throw new Sequelize.DatabaseError(error);
-    }
-
-}
-
-const getAPIId = async (orgID, apiHandle) => {
-
-    try {
-        const api = await APIMetadata.findOne({
-            attributes: ['API_ID'],
-            where: {
-                API_HANDLE: apiHandle,
-                ORG_ID: orgID
-            }
-        })
-        return api?.API_ID;
-    } catch (error) {
-        if (error instanceof Sequelize.EmptyResultError) {
-            throw error;
-        }
-        throw new Sequelize.DatabaseError(error);
-    }
-}
-
-const getAPIHandle = async (orgID, apiRefID) => {
-    try {
-        const api = await APIMetadata.findOne({
-            attributes: ['API_HANDLE'],
-            where: {
-                REFERENCE_ID: apiRefID,
-                ORG_ID: orgID
-            }
-        })
-        return api.API_HANDLE;
-    } catch (error) {
-        if (error instanceof Sequelize.EmptyResultError) {
-            throw error;
-        }
-        throw new Sequelize.DatabaseError(error);
-    }
-}
-
-const getApiIdByReferenceId = async (orgID, referenceId, t) => {
-    try {
-        const api = await APIMetadata.findOne({
-            attributes: ['API_ID'],
-            where: {
-                REFERENCE_ID: referenceId,
-                ORG_ID: orgID
-            },
-            transaction: t
-        });
-        return api?.API_ID;
-    } catch (error) {
-        if (error instanceof Sequelize.EmptyResultError) {
-            throw error;
-        }
-        throw new Sequelize.DatabaseError(error);
-    }
-};
-
-module.exports = {
-    createAPIMetadata,
-    createAPISubscriptionPolicy,
-    storeAPIFile,
-    getAPIMetadata,
-    getAllAPIMetadata,
-    storeAPIImageMetadata,
-    deleteAPIMetadata,
-    updateAPIMetadata,
-    updateAPISubscriptionPolicy,
-    updateAPIImageMetadata,
-    updateAPIFile,
-    storeAPIFiles,
-    updateOrCreateAPIFiles,
-    getAPIFile,
-    getAPIFileByType,
-    upsertAPIFileByType,
-    getAPIDoc,
-    deleteAPIFile,
-    deleteAllAPIFiles,
-    getAPIDocs,
-    getAPIDocLinks,
-    getAPIDocByName,
-    getAPIDocTypes,
-    getAPIId,
-    getAPIHandle,
-    getAPISpecs,
-    getAPIMetadataByCondition,
-    searchAPIMetadata,
-    putSubscriptionPolicy,
-    createSubscriptionPolicy,
-    bulkCreateSubscriptionPolicies,
-    getSubscriptionPolicyByName,
-    deleteSubscriptionPolicyById,
-    getSubscriptionPolicy,
-    getSubscriptionPolicies,
-    deleteSubscriptionPolicy,
-    createLabels,
-    getLabelID,
-    deleteLabel,
-    getLabels,
-    addView,
-    addViewLabels,
-    replaceViewLabels,
-    deleteViewLabels,
-    updateView,
-    deleteView,
-    getView,
-    getAllViews,
-    getViewID,
-    createAPILabelMapping,
-    deleteAPILabels,
-    updateLabel,
-    addLabel,
-    getImage,
-    deleteImage,
-    getAllSubscriptionPolicies,
-    getAllAPIMetadataFromAllViews,
-    getApiIdByReferenceId
-};
