+++ portals/developer-portal/src/utils/util.js
@@ -22,7 +22,7 @@ const marked = require('marked');
 const Handlebars = require('handlebars');
 const logger = require('../config/logger');
 const { CustomError } = require('../utils/errors/customErrors');
-const adminDao = require('../dao/adminDao');
+const orgDao = require('../dao/organizationDao');
 const constants = require('../utils/constants');
 const unzipper = require('unzipper');
 const axios = require('axios');
@@ -31,7 +31,8 @@ const https = require('https');
 const { config } = require('../config/configLoader');
 const { body, param, query } = require('express-validator');
 const { Sequelize } = require('sequelize');
-const apiDao = require('../dao/apiMetadataDao');
+const apiDao = require('../dao/apiDao');
+const subscriptionPolicyDao = require('../dao/subscriptionPolicyDao');
 const subscriptionPolicyDTO = require('../dto/subscriptionPolicyDto');
 const jwt = require('jsonwebtoken');
 const filePrefix = '/src/defaultContent/';
@@ -100,7 +101,7 @@ function renderTemplate(templatePath, layoutPath, templateContent, isTechnical) 
 
 async function loadLayoutFromAPI(orgID, viewName) {
 
-    var layoutContent = await adminDao.getOrgContent({
+    var layoutContent = await orgDao.getContent({
         orgId: orgID,
         fileType: constants.FILE_TYPE.LAYOUT,
         fileName: constants.FILE_NAME.MAIN,
@@ -115,7 +116,7 @@ async function loadLayoutFromAPI(orgID, viewName) {
 
 async function loadTemplateFromAPI(orgID, filePath, viewName) {
 
-    var templateContent = await adminDao.getOrgContent({
+    var templateContent = await orgDao.getContent({
         orgId: orgID,
         filePath: filePath,
         fileType: constants.FILE_TYPE.TEMPLATE,
@@ -131,7 +132,7 @@ async function renderTemplateFromAPI(templateContent, orgID, orgName, filePath, 
     const completeLayoutPath = path.join(process.cwd(), filePrefix + 'layout/main.hbs');
 
     layoutResponse = fs.readFileSync(completeLayoutPath, constants.CHARSET_UTF8);
-    const styleContent = await adminDao.getOrgContent({ orgId: orgID, fileType: 'style', viewName: viewName, fileName: 'main.css' });
+    const styleContent = await orgDao.getContent({ orgId: orgID, fileType: 'style', viewName: viewName, fileName: 'main.css' });
     if (styleContent) {
         layoutResponse = layoutResponse.replace(/\/styles\//g, `${constants.DEVPORTAL_API.orgPath(orgID)}/views/${viewName}/layout?fileType=style&fileName=`);
     }
@@ -156,7 +157,7 @@ async function renderTemplateFromAPI(templateContent, orgID, orgName, filePath, 
 
 async function renderLlmsTxt(templateContent, orgID, viewName) {
 
-    const dbPartial = await adminDao.getOrgContent({
+    const dbPartial = await orgDao.getContent({
         orgId: orgID,
         fileType: 'partial',
         viewName: viewName,
@@ -180,7 +181,7 @@ async function renderLlmsTxt(templateContent, orgID, viewName) {
 async function renderMarkdownTemplateFromAPI(templateContent, orgID, filePath, viewName) {
 
     const partialName = path.basename(filePath) + '-md';
-    const dbPartial = await adminDao.getOrgContent({
+    const dbPartial = await orgDao.getContent({
         orgId: orgID,
         fileType: 'partial',
         viewName: viewName,
@@ -796,7 +797,7 @@ async function appendSubscriptionPlanDetails(orgID, subscriptionPolicies) {
 const loadSubscriptionPlan = async (orgID, policyName) => {
 
     try {
-        const policyData = await apiDao.getSubscriptionPolicyByName(orgID, policyName);
+        const policyData = await subscriptionPolicyDao.getByName(orgID, policyName);
         if (policyData) {
             return new subscriptionPolicyDTO(policyData);
         } else {
@@ -821,7 +822,7 @@ async function tokenExchanger(token, orgName) {
     const url = config.advanced.tokenExchanger.url;
     const maxRetries = 3;
     let delay = 1000;
-    const orgDetails = await adminDao.getOrganization(orgName);
+    const orgDetails = await orgDao.get(orgName);
     if (!orgDetails) {
         throw new Error('Organization not found');
     } else if (!orgDetails.ORGANIZATION_IDENTIFIER) {
@@ -956,7 +957,7 @@ function filterAllowedAPIs(searchResults, allowedAPIs) {
 }
 
 const enforcePortalMode = async (req, res, next) => {
-    const orgDetails = await adminDao.getOrganization(req.params.orgName);
+    const orgDetails = await orgDao.get(req.params.orgName);
     const portalMode = orgDetails.ORG_CONFIG?.devportalMode || constants.DEVPORTAL_MODE.DEFAULT;
     const path = req.originalUrl.split('/')[4];
 
@@ -975,7 +976,7 @@ const enforcePortalMode = async (req, res, next) => {
 }
 
 async function isAiDisabledForPortal(orgID, viewName) {
-    const configAsset = await adminDao.getOrgContent({
+    const configAsset = await orgDao.getContent({
         orgId: orgID, fileType: constants.FILE_TYPE.LLMS_CONFIG, viewName, fileName: constants.FILE_NAME.LLMS_CONFIG
     });
     if (!configAsset) return false;
